--- examples/gpt/args.py
@@ -1,395 +1,0 @@
-# Copyright (c) 2022 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import argparse
-import paddle
-
-
-def process_batch_size(args):
-    if args.global_batch_size is None and args.local_batch_size is None:
-        raise ValueError(
-            "global_batch_size or local_batch_size should be set.")
-    elif args.global_batch_size is not None and args.local_batch_size is not None:
-        assert args.global_batch_size // args.local_batch_size == (args.dp_degree *
-            args.sharding_degree), "global_batch_size[{}] should be divided by local_batch_size[{}] "\
-            "when dp_degree is [{}] and sharding_degree is [{}]".format(args.global_batch_size,
-            args.local_batch_size, args.dp_degree, args.sharding_degree)
-    elif args.global_batch_size is not None and args.local_batch_size is None:
-        assert args.global_batch_size % (args.dp_degree * args.sharding_degree) == 0, \
-            "global_batch_size[{}] should be divided by dp_degree[{}] times sharding_degree[{}]"\
-            .format(args.global_batch_size, args.dp_degree, args.sharding_degree)
-        args.local_batch_size = args.global_batch_size // (
-            args.dp_degree * args.sharding_degree)
-    else:
-        args.global_batch_size = args.local_batch_size * args.dp_degree * args.sharding_degree
-    assert args.local_batch_size % args.micro_batch_size == 0
-
-
-def str2bool(v):
-    if v.lower() in ('yes', 'true', 't', 'y', '1'):
-        return True
-    elif v.lower() in ('no', 'false', 'f', 'n', '0'):
-        return False
-    else:
-        raise argparse.ArgumentTypeError('Unsupported value encountered.')
-
-
-def _add_training_args(parser):
-    group = parser.add_argument_group(title='training')
-
-    group.add_argument(
-        "--max_seq_len", type=int, default=1024, help="Max sequence length.")
-
-    group.add_argument(
-        "--global_batch_size",
-        default=None,
-        type=int,
-        help="Global batch size for all training process. None for not check the size is valid. If we only use data parallelism, it should be device_num * micro_batch_size."
-    )
-
-    group.add_argument(
-        "--local_batch_size",
-        default=None,
-        type=int,
-        help="Global batch size for all training process. None for not check the size is valid. If we only use data parallelism, it should be device_num * micro_batch_size."
-    )
-
-    group.add_argument(
-        "--micro_batch_size",
-        default=8,
-        type=int,
-        help="Batch size per device for one step training.", )
-
-    return parser
-
-
-def _add_model_size(parser):
-    group = parser.add_argument_group(title='network_size')
-
-    group.add_argument(
-        '--vocab_size', type=int, default=51200, help='Vocab size')
-    group.add_argument(
-        '--hidden_size',
-        type=int,
-        default=1024,
-        help='Tansformer hidden size.')
-    group.add_argument(
-        '--num_layers',
-        type=int,
-        default=12,
-        help='Number of transformer layers.')
-
-    group.add_argument(
-        '--num_attention_heads',
-        type=int,
-        default=12,
-        help='Number of transformer attention heads.')
-
-    group.add_argument(
-        '--ffn_hidden_size',
-        type=int,
-        default=None,
-        help='Transformer Feed-Forward Network hidden size. '
-        'This is set to 4*hidden-size if not provided')
-
-    group.add_argument(
-        "--hidden_dropout_prob",
-        type=float,
-        default=0.1,
-        help="The hidden dropout prob.")
-    group.add_argument(
-        "--attention_probs_dropout_prob",
-        type=float,
-        default=0.1,
-        help="The attention probs dropout prob.")
-
-    group.add_argument(
-        '--max_position_embeddings',
-        type=int,
-        default=1024,
-        help='Maximum number of position embeddings to use. '
-        'This is the size of position embedding.')
-
-    group.add_argument(
-        '--type_vocab_size', type=int, default=16, help='type_vocab_size')
-
-    group.add_argument(
-        '--initializer_range',
-        type=float,
-        default=0.02,
-        help='initializer_range')
-
-    return parser
-
-
-def _add_data_args(parser):
-    group = parser.add_argument_group(title='data arguments')
-    # Train I/O config
-    group.add_argument(
-        "--input_dir",
-        default=None,
-        type=str,
-        required=True,
-        help="The input directory where the data will be read from.", )
-    group.add_argument(
-        "--output_dir",
-        default=None,
-        type=str,
-        required=True,
-        help="The output directory where the training logs and checkpoints will be written."
-    )
-    group.add_argument(
-        "--split",
-        type=str,
-        default='949,50,1',
-        help="Train/valid/test data split.")
-
-    group.add_argument(
-        "--ckpt_dir",
-        default=None,
-        type=str,
-        required=False,
-        help="The checkpoint directory.")
-
-    return parser
-
-
-def _add_training_steps_args(parser):
-    group = parser.add_argument_group(title='training steps')
-
-    group.add_argument(
-        "--decay_steps",
-        default=360000,
-        type=int,
-        help="The steps use to control the learing rate. If the step > decay_steps, will use the min_lr."
-    )
-    group.add_argument(
-        "--max_steps",
-        default=500000,
-        type=int,
-        help="If > 0: set total number of training steps to perform. Override num_train_epochs."
-    )
-
-    group.add_argument(
-        "--warmup_rate",
-        default=0.01,
-        type=float,
-        help="Linear warmup over warmup_steps for learing rate.")
-
-    group.add_argument(
-        "--max_lr",
-        default=1e-5,
-        type=float,
-        help="The initial max learning rate for Adam.")
-    group.add_argument(
-        "--min_lr",
-        default=5e-5,
-        type=float,
-        help="The initial min learning rate for Adam.")
-    group.add_argument(
-        "--grad_clip",
-        default=0.0,
-        type=float,
-        help="Grad clip for the parameter.")
-
-    return parser
-
-
-def _add_optimizer_args(parser):
-    group = parser.add_argument_group(title='optimizer arguments')
-    group.add_argument(
-        "--weight_decay",
-        default=0.0,
-        type=float,
-        help="Weight decay if we apply some.")
-
-    # Adam optimizer config
-    group.add_argument(
-        "--adam_beta1",
-        default=0.9,
-        type=float,
-        help="The beta1 for Adam optimizer. The exponential decay rate for the 1st moment estimates."
-    )
-    group.add_argument(
-        "--adam_beta2",
-        default=0.999,
-        type=float,
-        help="The bate2 for Adam optimizer. The exponential decay rate for the 2nd moment estimates."
-    )
-    group.add_argument(
-        "--adam_epsilon",
-        default=1e-8,
-        type=float,
-        help="Epsilon for Adam optimizer.")
-    return parser
-
-
-def _add_fp16_args(parser):
-    group = parser.add_argument_group(title='fp16 arguments')
-    group.add_argument(
-        "--use_pure_fp16",
-        type=str2bool,
-        nargs='?',
-        const=False,
-        help="Enable pure fp16 precision training.")
-
-    group.add_argument(
-        "--scale_loss",
-        type=float,
-        default=32768,
-        help="The value of scale_loss for fp16. This is only used for AMP training."
-    )
-
-    return parser
-
-
-def _add_training_step(parser):
-    group = parser.add_argument_group(title='training step')
-    group.add_argument(
-        "--num_train_epochs",
-        default=1,
-        type=int,
-        help="Total number of training epochs to perform.", )
-    group.add_argument(
-        "--logging_freq",
-        type=int,
-        default=1,
-        help="Log every X updates steps.")
-
-    group.add_argument(
-        "--eval_freq",
-        type=int,
-        default=500,
-        help="Evaluate for every X updates steps.")
-
-    group.add_argument(
-        "--eval_iters",
-        type=int,
-        default=10,
-        help="Evaluate the model use X steps data.")
-
-    parser.add_argument(
-        "--save_steps",
-        type=int,
-        default=500,
-        help="Save checkpoint every X updates steps.")
-
-    return parser
-
-
-def _add_distributed_args(parser):
-    group = parser.add_argument_group(title='distributed arguments')
-
-    # Config for 4D Parallelism
-    group.add_argument(
-        "--sharding_degree",
-        type=int,
-        default=1,
-        help="Sharding degree. Share the parameters to many cards.")
-
-    group.add_argument(
-        "--dp_degree", type=int, default=1, help="Data Parallelism degree.")
-
-    group.add_argument(
-        "--mp_degree",
-        type=int,
-        default=1,
-        help="Model Parallelism degree. Spliting the linear layers to many cards."
-    )
-
-    group.add_argument(
-        "--pp_degree",
-        type=int,
-        default=1,
-        help="Pipeline Parallelism degree.  Spliting the the model layers to different parts."
-    )
-
-    group.add_argument(
-        "--use_recompute",
-        type=str2bool,
-        nargs='?',
-        const=False,
-        help="Using the recompute to save the memory.")
-
-    # add sharding stage2/3
-    group.add_argument(
-        "--sharding_stage",
-        type=int,
-        default=1,
-        help="sharding stage1/2/3. Stage 1: The optimizer states are partitioned across the processes, so that each process updates only its partition. Stage 2: The reduced gradients for updating the model weights are also partitioned such that each process retains only the gradients corresponding to its portion of the optimizer states. Stage 3: The model parameters are partitioned across the processes. stage3 will automatically collect and partition them during the forward and backward passes."
-    )
-
-    group.add_argument(
-        "--sharding_offload",
-        type=str2bool,
-        nargs='?',
-        const=False,
-        help="sharding stage2/3 cpu offload strategy.")
-
-    return parser
-
-
-def parse_args():
-    parser = argparse.ArgumentParser()
-
-    parser.add_argument(
-        "--device",
-        type=str,
-        default="gpu",
-        choices=["cpu", "gpu", "xpu"],
-        help="select cpu, gpu, xpu devices.")
-
-    parser = _add_training_args(parser)
-    parser = _add_model_size(parser)
-    parser = _add_data_args(parser)
-    parser = _add_training_steps_args(parser)
-    parser = _add_optimizer_args(parser)
-    parser = _add_fp16_args(parser)
-    parser = _add_training_step(parser)
-    parser = _add_distributed_args(parser)
-
-    # Other config
-    parser.add_argument(
-        "--seed",
-        type=int,
-        default=1234,
-        help="Random seed for initialization")
-
-    args = parser.parse_args()
-    args.test_iters = args.eval_iters * 10
-
-    # process batch size
-    process_batch_size(args)
-
-    if args.ffn_hidden_size is None:
-        args.ffn_hidden_size = 4 * args.hidden_size
-
-    _print_args(args)
-    return args
-
-
-def _print_args(args):
-    """Print arguments."""
-    print(
-        '------------------------ arguments ------------------------',
-        flush=True)
-    str_list = []
-    for arg in vars(args):
-        dots = '.' * (48 - len(arg))
-        str_list.append('  {} {} {}'.format(arg, dots, getattr(args, arg)))
-    for arg in sorted(str_list, key=lambda x: x.lower()):
-        print(arg, flush=True)
-    print(
-        '-------------------- end of arguments ---------------------',
-        flush=True)
